moving metadataquery
