+++ lib/private/FilesMetadata/FilesMetadataManager.php
@@ -30,7 +30,6 @@ use OC\FilesMetadata\Job\UpdateSingleMetadata;
 use OC\FilesMetadata\Listener\MetadataDelete;
 use OC\FilesMetadata\Listener\MetadataUpdate;
 use OC\FilesMetadata\Model\FilesMetadata;
-use OC\FilesMetadata\Model\MetadataQuery;
 use OC\FilesMetadata\Service\IndexRequestService;
 use OC\FilesMetadata\Service\MetadataRequestService;
 use OCP\BackgroundJob\IJobList;
@@ -48,8 +47,8 @@ use OCP\FilesMetadata\Event\MetadataLiveEvent;
 use OCP\FilesMetadata\Exceptions\FilesMetadataException;
 use OCP\FilesMetadata\Exceptions\FilesMetadataNotFoundException;
 use OCP\FilesMetadata\IFilesMetadataManager;
+use OCP\FilesMetadata\IMetadataQuery;
 use OCP\FilesMetadata\Model\IFilesMetadata;
-use OCP\FilesMetadata\Model\IMetadataQuery;
 use OCP\FilesMetadata\Model\IMetadataValueWrapper;
 use OCP\IConfig;
 use Psr\Log\LoggerInterface;
